requirements.txt
langchain==0.3.26
flask==3.1.1 
sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
faiss-cpu==1.8.1
langchain_groq==0.3.16 
langchain-openai==0.3.24
langchain-community==0.3.26
-e .